requirements.txt
#langchain==0.0.259
langchain==0.0.258
langchainplus-sdk==0.0.19
openai==0.27.8
openapi-schema-pydantic==1.2.4
pydantic==1.10.10
python-dotenv
langcorn==0.0.13
